Let’s do it the right way this time 👏
Musixmatch wants to be annoying, so they add a few lines of the lyrics in a separate container under the same properties. Additionally, there are some cases (see: musixmatch.com/lyrics/Vijay-feat-Sofia-Zlatko-Tania-Zygar/Wildest-Dreams) where the lyrics are fine/valid, and public, but on “review” for months or more. This fixes (mostly for the latter) and finds all the matching DOM elements. In the first case, it’ll combine the “preview” lines properly, and in the second case it will properly display the lyrics instead of just seeing the “Review” message as the lyrics. Since this property is exclusive to actual lyric content, it’s safe to work. 🙃